reformat codebase

+++ src/middlewares/hot-reload.js
@@ -16,7 +16,6 @@ const { clearRequireCache, makeDirs, isFunction, isString } = require("../utils"
 
 /* istanbul ignore next */
 module.exports = function HotReloadMiddleware(broker) {
-
 	const cache = new Map();
 
 	let projectFiles = new Map();
@@ -29,13 +28,11 @@ module.exports = function HotReloadMiddleware(broker) {
 
 		broker.logger.info(`Hot reload '${service.name}' service...`, kleur.grey(relPath));
 
-		return broker.destroyService(service)
-			.then(() => {
-				if (fs.existsSync(service.__filename)) {
-					return broker.loadService(service.__filename);
-				}
-			});
-
+		return broker.destroyService(service).then(() => {
+			if (fs.existsSync(service.__filename)) {
+				return broker.loadService(service.__filename);
+			}
+		});
 	}
 
 	/**
@@ -43,7 +40,6 @@ module.exports = function HotReloadMiddleware(broker) {
 	 *
 	 */
 	function watchProjectFiles() {
-
 		if (!broker.started || !process.mainModule) return;
 
 		cache.clear();
@@ -59,14 +55,10 @@ module.exports = function HotReloadMiddleware(broker) {
 		if (extraFiles != null) {
 			Object.entries(extraFiles).forEach(([fName, restartType]) => {
 				const watchItem = getWatchItem(fName);
-				if (restartType == "broker")
-					watchItem.brokerRestart = true;
-				else if (restartType == "allServices")
-					watchItem.allServices = true;
-				else if (isString(restartType))
-					watchItem.services.push(restartType);
-				else if (Array.isArray(restartType))
-					watchItem.services.push(...restartType);
+				if (restartType == "broker") watchItem.brokerRestart = true;
+				else if (restartType == "allServices") watchItem.allServices = true;
+				else if (isString(restartType)) watchItem.services.push(restartType);
+				else if (Array.isArray(restartType)) watchItem.services.push(...restartType);
 			});
 		}
 
@@ -80,19 +72,18 @@ module.exports = function HotReloadMiddleware(broker) {
 				return ac == bc;
 			});
 
-			broker.logger.info(kleur.bgMagenta().white().bold(`Reload ${needToReloadDedup.length} service(s)`));
+			broker.logger.info(
+				kleur.bgMagenta().white().bold(`Reload ${needToReloadDedup.length} service(s)`)
+			);
 
 			needToReloadDedup.forEach(svc => {
 				if (typeof svc == "string")
-					if (fs.existsSync(svc))
-						return broker.loadService(svc);
-					else
-						return;
+					if (fs.existsSync(svc)) return broker.loadService(svc);
+					else return;
 
 				return hotReloadService(svc);
 			});
 			needToReload.clear();
-
 		}, 500);
 
 		// Close previous watchers
@@ -104,8 +95,7 @@ module.exports = function HotReloadMiddleware(broker) {
 
 		projectFiles.forEach((watchItem, fName) => {
 			// Delete if file doesn't exist anymore
-			if (!fs.existsSync(fName))
-				projectFiles.delete(fName);
+			if (!fs.existsSync(fName)) projectFiles.delete(fName);
 		});
 
 		projectFiles.forEach((watchItem, fName) => {
@@ -115,14 +105,25 @@ module.exports = function HotReloadMiddleware(broker) {
 			else if (watchItem.allServices)
 				broker.logger.debug(`  ${relPath}:`, kleur.grey("reload all services."));
 			else if (watchItem.services.length > 0) {
-				broker.logger.debug(`  ${relPath}:`, kleur.grey(`reload ${watchItem.services.length} service(s) & ${watchItem.others.length} other(s).`)/*, watchItem.services, watchItem.others*/);
-				watchItem.services.forEach(svcFullname => broker.logger.debug(kleur.grey(`    ${svcFullname}`)));
-				watchItem.others.forEach(filename => broker.logger.debug(kleur.grey(`    ${path.relative(process.cwd(), filename)}`)));
+				broker.logger.debug(
+					`  ${relPath}:`,
+					kleur.grey(
+						`reload ${watchItem.services.length} service(s) & ${watchItem.others.length} other(s).`
+					) /*, watchItem.services, watchItem.others*/
+				);
+				watchItem.services.forEach(svcFullname =>
+					broker.logger.debug(kleur.grey(`    ${svcFullname}`))
+				);
+				watchItem.others.forEach(filename =>
+					broker.logger.debug(kleur.grey(`    ${path.relative(process.cwd(), filename)}`))
+				);
 			}
 			// Create watcher
-			watchItem.watcher = fs.watch(fName, (eventType) => {
+			watchItem.watcher = fs.watch(fName, eventType => {
 				const relPath = path.relative(process.cwd(), fName);
-				broker.logger.info(kleur.magenta().bold(`The '${relPath}' file is changed. (Event: ${eventType})`));
+				broker.logger.info(
+					kleur.magenta().bold(`The '${relPath}' file is changed. (Event: ${eventType})`)
+				);
 
 				// Clear from require cache
 				clearRequireCache(fName);
@@ -130,26 +131,26 @@ module.exports = function HotReloadMiddleware(broker) {
 					watchItem.others.forEach(f => clearRequireCache(f));
 				}
 
-				if (watchItem.brokerRestart && broker.runner && isFunction(broker.runner.restartBroker)) {
+				if (
+					watchItem.brokerRestart &&
+					broker.runner &&
+					isFunction(broker.runner.restartBroker)
+				) {
 					broker.logger.info(kleur.bgMagenta().white().bold("Action: Restart broker..."));
 					stopAllFileWatcher(projectFiles);
 					// Clear the whole require cache
 					require.cache.length = 0;
 					return broker.runner.restartBroker();
-
 				} else if (watchItem.allServices) {
 					// Reload all services
 					broker.services.forEach(svc => {
-						if (svc.__filename)
-							needToReload.add(svc);
+						if (svc.__filename) needToReload.add(svc);
 					});
 					reloadServices();
-
 				} else if (watchItem.services.length > 0) {
 					// Reload certain services
 					broker.services.forEach(svc => {
-						if (watchItem.services.indexOf(svc.fullName) !== -1)
-							needToReload.add(svc);
+						if (watchItem.services.indexOf(svc.fullName) !== -1) needToReload.add(svc);
 					});
 
 					if (needToReload.size == 0) {
@@ -164,9 +165,7 @@ module.exports = function HotReloadMiddleware(broker) {
 			});
 		});
 
-		if (projectFiles.size == 0)
-			broker.logger.debug(kleur.grey("  No files."));
-
+		if (projectFiles.size == 0) broker.logger.debug(kleur.grey("  No files."));
 	}
 
 	const debouncedWatchProjectFiles = _.debounce(watchProjectFiles, 2000);
@@ -175,7 +174,7 @@ module.exports = function HotReloadMiddleware(broker) {
 	 * Stop all file watchers
 	 */
 	function stopAllFileWatcher(items) {
-		items.forEach((watchItem) => {
+		items.forEach(watchItem => {
 			if (watchItem.watcher) {
 				watchItem.watcher.close();
 				watchItem.watcher = null;
@@ -191,8 +190,7 @@ module.exports = function HotReloadMiddleware(broker) {
 	 */
 	function getWatchItem(fName) {
 		let watchItem = projectFiles.get(fName);
-		if (watchItem)
-			return watchItem;
+		if (watchItem) return watchItem;
 
 		watchItem = {
 			services: [],
@@ -206,9 +204,11 @@ module.exports = function HotReloadMiddleware(broker) {
 	}
 
 	function isMoleculerConfig(fName) {
-		return fName.endsWith("moleculer.config.js")
-			|| fName.endsWith("moleculer.config.ts")
-			|| fName.endsWith("moleculer.config.json");
+		return (
+			fName.endsWith("moleculer.config.js") ||
+			fName.endsWith("moleculer.config.ts") ||
+			fName.endsWith("moleculer.config.json")
+		);
 	}
 
 	/**
@@ -227,8 +227,7 @@ module.exports = function HotReloadMiddleware(broker) {
 				return;
 
 		// Avoid circular dependency in project files
-		if (parents && parents.indexOf(fName) !== -1)
-			return;
+		if (parents && parents.indexOf(fName) !== -1) return;
 
 		// console.log(fName);
 
@@ -249,7 +248,6 @@ module.exports = function HotReloadMiddleware(broker) {
 				watchItem.services.push(service.fullName);
 
 			watchItem.others = _.uniq([].concat(watchItem.others, parents || []));
-
 		} else if (isMoleculerConfig(fName)) {
 			const watchItem = getWatchItem(fName);
 			watchItem.brokerRestart = true;
@@ -282,20 +280,20 @@ module.exports = function HotReloadMiddleware(broker) {
 		// Debounced Service loader function
 		const needToLoad = new Set();
 		const loadServices = _.debounce(() => {
-			broker.logger.info(kleur.bgMagenta().white().bold(`Load ${needToLoad.size} service(s)...`));
+			broker.logger.info(
+				kleur.bgMagenta().white().bold(`Load ${needToLoad.size} service(s)...`)
+			);
 
 			needToLoad.forEach(filename => {
 				try {
 					broker.loadService(filename);
-				} catch(err) {
+				} catch (err) {
 					broker.logger.error(`Failed to load service '${filename}'`, err);
 					clearRequireCache(filename);
 				}
 			});
 			needToLoad.clear();
-
 		}, 500);
-
 
 		if (broker.runner && Array.isArray(broker.runner.folders)) {
 			const folders = broker.runner.folders;
@@ -310,14 +308,25 @@ module.exports = function HotReloadMiddleware(broker) {
 					broker.logger.debug(`  ${path.relative(process.cwd(), folder)}/`);
 					folderWatchers.push({
 						path: folder,
-						watcher: watch(folder, (filename) => {
-							if (filename.endsWith(".service.js") || filename.endsWith(".service.ts")) {
-								broker.logger.debug(`There is changes in '${folder}' folder: `, path.basename(filename));
-								const isLoaded = broker.services.some(svc => svc.__filename == filename);
+						watcher: watch(folder, filename => {
+							if (
+								filename.endsWith(".service.js") ||
+								filename.endsWith(".service.ts")
+							) {
+								broker.logger.debug(
+									`There is changes in '${folder}' folder: `,
+									path.basename(filename)
+								);
+								const isLoaded = broker.services.some(
+									svc => svc.__filename == filename
+								);
 								const fileExists = fs.existsSync(filename);
 								if (!isLoaded && fileExists) {
 									// This is a new file. We should wait for the file fully copied.
-									broker.logger.debug("Loading new file: ", path.basename(filename));
+									broker.logger.debug(
+										"Loading new file: ",
+										path.basename(filename)
+									);
 									needToLoad.add(filename);
 									loadServices();
 								}
@@ -347,7 +356,9 @@ module.exports = function HotReloadMiddleware(broker) {
 				return;
 			} else if (typeof broker.options.hotReload === "object") {
 				if (Array.isArray(broker.options.hotReload.modules)) {
-					hotReloadModules = broker.options.hotReload.modules.map(moduleName => `/node_modules/${moduleName}/`);
+					hotReloadModules = broker.options.hotReload.modules.map(
+						moduleName => `/node_modules/${moduleName}/`
+					);
 				}
 				if (broker.options.hotReload.extraFiles) {
 					/**
@@ -385,5 +396,4 @@ module.exports = function HotReloadMiddleware(broker) {
 			stopProjectFolderWatchers();
 		}
 	};
-
 };
